feat: implement baseline for home
