+++ src/layout/cube.jsx
@@ -1,0 +1,17 @@
+export default function Cube({ width = 418, height = 571 }) {
+  return (
+    <svg
+      width={width}
+      height={height}
+      viewBox={`0 0 ${width} ${height}`}
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+    >
+      <path
+        d="M4.3539 513.244H363.587L414.614 569.5H76.3088L4.3539 513.244ZM362.751 510.244H1.5V1.5H362.751V510.244ZM365.751 511.165V4.39872L416.5 70.1559V567.114L365.751 511.165Z"
+        stroke="white"
+        stroke-width="3"
+      />
+    </svg>
+  )
+}
